moved find_center() to prep.py; improved log info
